requirements.txt
Flask==3.1.3
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
Werkzeug==3.1.6
flask-cors==6.0.2
python-dotenv==1.2.2
